separated randomize function
